requirements.txt
streamlit>=1.28.0
supabase>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
beautifulsoup4>=4.12.0
fpdf2>=2.7.0
nest-asyncio>=1.5.0
